backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
google-genai==0.2.2
python-dotenv==1.0.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
bleach==6.1.0
